Rename PermissionsSelector
